Move conda service to correct module
